refactor(eve): own subagent workflow tools in framework
