+++ packages/eve/src/execution/tasks/parent/subagent/local.ts
@@ -1,4 +1,4 @@
-import { z } from "#compiled/zod/index.js";
+import type { z } from "#compiled/zod/index.js";
 import type {
   DispatchPlanEntry,
   DispatchStartTarget,
@@ -6,10 +6,8 @@ import type {
   RuntimeActionDispatchResult,
 } from "#execution/dispatch-runtime-actions-shared.js";
 import { dispatchSubagentWorkflowToolStep } from "#execution/tasks/parent/subagent/dispatch-step.js";
-import type { ToolDefinition } from "#public/definitions/tool.js";
-import { defineTool } from "#public/definitions/tool.js";
 import type { RuntimeSubagentCallActionRequest } from "#runtime/actions/types.js";
-import { PERSISTENT_SUBAGENT_TOOL_INPUT_SCHEMA } from "#runtime/subagents/registry.js";
+import type { PERSISTENT_SUBAGENT_TOOL_INPUT_SCHEMA } from "#runtime/subagents/registry.js";
 
 type ResumeEntry = Extract<DispatchPlanEntry, { readonly kind: "resume" }>;
 type StartEntry = Extract<DispatchPlanEntry, { readonly kind: "start" }>;
@@ -45,20 +43,6 @@ export async function executeLocalSubagentWorkflow(
     transport: "local",
   });
 }
-
-export type LocalSubagentWorkflowTool = ToolDefinition<
-  LocalSubagentWorkflowInput,
-  RuntimeActionDispatchResult
->;
-
-// Public tools receive a live ToolContext. This framework-owned Workflow is
-// started by workflowId with a serializable private context.
-export const localSubagentWorkflowTool: LocalSubagentWorkflowTool = defineTool({
-  description: "Dispatch a local subagent as a durable background task.",
-  execute: executeLocalSubagentWorkflow as LocalSubagentWorkflowTool["execute"] &
-    typeof executeLocalSubagentWorkflow,
-  inputSchema: PERSISTENT_SUBAGENT_TOOL_INPUT_SCHEMA,
-});
 
 export function isLocalSubagentWorkflowEntry(
   entry: Extract<DispatchPlanEntry, { readonly kind: "resume" | "start" }>,
