+++ packages/eve/src/runtime/resolve-agent-graph.ts
@@ -26,10 +26,7 @@ import { resolveDynamicSubagentDefinition } from "#runtime/resolve-dynamic-subag
 import { loadResolvedModuleExport } from "#runtime/resolve-helpers.js";
 import { createRuntimeSandboxRegistry } from "#runtime/sandbox/registry.js";
 import { LOAD_SKILL_TOOL_NAME } from "#runtime/skills/fragment-context.js";
-import {
-  createRuntimeSubagentRegistry,
-  type RuntimeSubagentTaskTools,
-} from "#runtime/subagents/registry.js";
+import { createRuntimeSubagentRegistry } from "#runtime/subagents/registry.js";
 import { createRuntimeToolRegistry } from "#runtime/tools/registry.js";
 import { WORKFLOW_TOOL_NAME } from "#shared/workflow-sandbox.js";
 import type {
@@ -235,8 +232,6 @@ async function resolveRuntimeAgentNode(
       parentNodeId: input.nodeId,
       subagentNodesById: input.subagentNodesById,
     }),
-    taskTools:
-      agent.config?.experimental?.tasks === true ? await loadSubagentTaskTools() : undefined,
   });
   const resolvedAgent = {
     ...agent,
@@ -263,19 +258,6 @@ async function resolveRuntimeAgentNode(
   input.nodesByNodeId.set(nodeId, node);
 
   return node;
-}
-
-async function loadSubagentTaskTools(): Promise<RuntimeSubagentTaskTools> {
-  // Keep the Workflow modules out of flag-off graph resolution. A static
-  // import would also cycle through their dispatch step back into this loader.
-  const [{ localSubagentWorkflowTool }, { remoteSubagentWorkflowTool }] = await Promise.all([
-    import("#execution/tasks/parent/subagent/local.js"),
-    import("#execution/tasks/parent/subagent/remote.js"),
-  ]);
-  return {
-    local: localSubagentWorkflowTool,
-    remote: remoteSubagentWorkflowTool,
-  };
 }
 
 async function resolveRuntimeSubagents(input: {
